requirements.txt
# ── Core ML / LLM ───────────────────────────────────
torch>=2.1.0
transformers>=4.40.0
accelerate>=0.27.0
sentence-transformers>=2.7.0

# ── Vector Store ────────────────────────────────────
chromadb>=0.5.0

# ── LangChain (Document abstraction only) ──────────
langchain-core>=0.2.0

# ── Datasets ────────────────────────────────────────
datasets>=2.19.0

# ── Evaluation Metrics ─────────────────────────────
rouge-score>=0.1.2

# ── GMM Chunker Dependencies ───────────────────────
scikit-learn>=1.4.0
numpy>=1.26.0
umap-learn>=0.5.0
spacy>=3.7.0

# ── Utilities ───────────────────────────────────────
tqdm>=4.66.0
tabulate>=0.9.0
nltk>=3.8.0
